Fix python binary link

+++ Dockerfile
@@ -48,7 +48,9 @@ EOF
 RUN <<EOF
     set -euxo pipefail
     export PYTHONUNBUFFERED=1
-    ln -s /usr/bin/python3 /usr/bin/python
+    if [ ! -f /usr/bin/python ]; then
+        ln -s /usr/bin/python3 /usr/bin/python
+    fi
     python3 -m ensurepip
     pip3 install --no-cache --upgrade pip
     pip3 install --no-cache unoserver
